refactor: split index style file into smaller ones
