+++ src/shared/styles/buttonStyle.ts
@@ -1,0 +1,35 @@
+import { StyleSheet } from 'react-native';
+
+export const buttonStyle = StyleSheet.create({
+  normal: {
+    backgroundColor: '#0066FF',
+    color: '#FFFFFF',
+    textAlign: 'center',
+    padding: 13,
+    borderRadius: 8,
+  },
+  white: {
+    backgroundColor: '#FFFFFF',
+    color: '#000000',
+    padding: 13,
+    borderRadius: 8,
+  },
+  black: {
+    backgroundColor: '#FFFFFF',
+    color: '#000000',
+    padding: 13,
+    borderRadius: 8,
+  },
+  gray: {
+    backgroundColor: '#FFFFFF',
+    color: '#000000',
+    padding: 13,
+    borderRadius: 8,
+  },
+  dark: {
+    backgroundColor: '#161616',
+    color: '#FFFFFF',
+    padding: 13,
+    borderRadius: 8,
+  },
+});
