+++ src/shared/styles/colors.ts
@@ -1,0 +1,11 @@
+export const styledColors = {
+  normal: '#0066FF',
+  black: '#000000',
+  dark: '#161616',
+  darkGray: '#303030',
+  gray: '#808080',
+  lightGray: '#D4D4D4',
+  light: '#EFEFEF',
+  white: '#FFFFFF',
+  border: '#D5D5D5',
+};
