+++ src/shared/styles/typography.ts
@@ -1,0 +1,52 @@
+import { StyleSheet } from 'react-native';
+
+export const fontStyle = StyleSheet.create({
+  fontH1: {
+    fontFamily: 'Clash Display Variable',
+    fontSize: 32,
+    fontWeight: '600',
+    lineHeight: 36,
+    letterSpacing: 0.05,
+    textAlign: 'center',
+  },
+  fontH2: {
+    fontFamily: 'Satoshi Variable',
+    fontSize: 24,
+    fontWeight: '600',
+    lineHeight: 32.5,
+    letterSpacing: 0,
+    textAlign: 'center',
+  },
+  fontH3: {
+    fontFamily: 'Satoshi Variable',
+    fontSize: 14,
+    fontWeight: '500',
+    lineHeight: 19,
+    letterSpacing: 0,
+    textAlign: 'center',
+  },
+  fontParagraph: {
+    fontFamily: 'Satoshi Variable',
+    fontSize: 14,
+    fontWeight: '400',
+    lineHeight: 19,
+    letterSpacing: 0,
+    textAlign: 'center',
+  },
+  fontSmallText: {
+    fontFamily: 'Satoshi Variable',
+    fontSize: 12,
+    fontWeight: '400',
+    lineHeight: 16,
+    letterSpacing: 0,
+    textAlign: 'center',
+  },
+  fontSmallTextBold: {
+    fontFamily: 'Satoshi Variable',
+    fontSize: 12,
+    fontWeight: '600',
+    lineHeight: 16,
+    letterSpacing: 0,
+    textAlign: 'left',
+  },
+});
